create a new file for all the promise logic
